% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annot.r
\name{annot}
\alias{annot}
\title{Annotation function for Symbol, NCBI or Ensembl IDs}
\usage{
annot(
  symbollist = NULL,
  species = NULL,
  ortholog = F,
  dboutput = "ncbi",
  idtype = NULL
)
}
\arguments{
\item{symbollist}{character list of IDs or Symbols}

\item{species}{character species 'hs' 'mm' 'rn' 'dr' 'ss' (see details for complete list)}

\item{ortholog}{logical if TRUE return homo sapiens ortholog for choosen species}

\item{dboutput}{character database used for output 'ncbi'(default) or 'ebi'}

\item{idtype}{character database ID accepted: 'SYMBOL'(default), 'GENE', 'ENST', 'ENSG', 'ENSP'}
}
\value{
data.frame
}
\description{
Annotation function for Symbol, NCBI or Ensembl IDs
}
\details{
Use moal:::orthoinfo to see complete species list
}
\examples{
# not run
# annot(Symbol)
}
\author{
Florent Dumont \href{mailto:florent.dumont@universite-paris-saclay.fr}{florent.dumont@universite-paris-saclay.fr}
}
