% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omic.r
\name{omic}
\alias{omic}
\title{Omic bioanalysis workflow}
\usage{
omic(
  dat = NULL,
  sif = NULL,
  annot = NULL,
  species = "hs",
  model = NULL,
  paired = NULL,
  nested = NULL,
  batch = NULL,
  addfactor = NULL,
  doqc = TRUE,
  threshold = c(1, 2, 3, 4, 9, 10, 11, 12),
  padj = "none",
  logratio = FALSE,
  dopattern = TRUE,
  dovenn = TRUE,
  docluster = TRUE,
  nc = c(2, 3, 6, 12),
  maxclusterheatmap = 5000,
  doheatmap = TRUE,
  heatmapcluster = "row",
  maxheatmap = 2000,
  minheatmap = 3,
  dovolcanoplot = TRUE,
  nbgenevolc = 5,
  dolineplot = TRUE,
  doboxplotrow = TRUE,
  doena = TRUE,
  gsearank = "logfc",
  gseatail = "twotail",
  topdeg = 100,
  topena = 50,
  doenaora = FALSE,
  gmtfiles = NULL,
  filtergeneset = NULL,
  bg = 25000,
  dotopnetwork = TRUE,
  dotopheatmap = TRUE,
  layout = 2,
  mings = 5,
  maxgs = 700,
  overlapmin = 2,
  addenarankbarplot = TRUE,
  dotopgenesetnetwork = FALSE,
  dotopgenesetheatmap = TRUE,
  dogmtgenesetnetwork = FALSE,
  dogmtgenesetheatmap = TRUE,
  crosscompint = FALSE,
  sample = NULL,
  seed = 123679,
  dopar = NULL,
  path = ".",
  dirname = NULL,
  zip = FALSE,
  remove = FALSE
)
}
\arguments{
\item{dat}{data.frame normalize data table with rowID for first column}

\item{sif}{data.frame sample information file including model factors}

\item{annot}{data.frame annotation with Symbol column for functional analysis}

\item{species}{character available species: bt ce dr dm gg hs mm pt rn ss xt}

\item{model}{character anova model factors (see details)}

\item{paired}{character factor for paired design}

\item{nested}{character factor for nested design}

\item{batch}{character factor for batch effect design}

\item{addfactor}{character additionnal factors}

\item{doqc}{logical quality controls}

\item{threshold}{numeric vector from 1 to 24 (see details)}

\item{padj}{character fdr by defaut for Benjamini-Hochberg false discovery correction}

\item{logratio}{logical change fc (by default) in log2ratio}

\item{dopattern}{logical search relevant pattern across levels factor}

\item{dovenn}{logical venn diagram}

\item{docluster}{logical row hierarchical clustering using pearson correlation}

\item{nc}{numeric number of clusters to cut in dendrogramm}

\item{maxclusterheatmap}{numeric max row for cluster analysis}

\item{doheatmap}{logical do heatmaps for all lists}

\item{heatmapcluster}{character row clustering only by default both accepted}

\item{maxheatmap}{numeric max rows for heatmap}

\item{minheatmap}{numeric min rows for heatmap}

\item{dovolcanoplot}{logical make volcanoplot for each threshold}

\item{nbgenevolc}{numeric number of Symbol to display in volcanoplot}

\item{dolineplot}{logical do lineplot for significant features}

\item{doboxplotrow}{logical do boxplot for significant features with Kruskal}

\item{doena}{logical msigdb enrichement analysis using gsea method without filtering}

\item{gsearank}{character to choose gsea rank type among fc (by default) logration logfc sqrt}

\item{gseatail}{character to choose gsea twotail (by default) or onetail}

\item{topdeg}{numeric top DEGs number to plot on network}

\item{topena}{numeric top geneset for ena plot}

\item{doenaora}{logical msigdb enrichement analysis using ora method for diff list}

\item{gmtfiles}{character gmt files list path}

\item{filtergeneset}{character regular expression to filter collection geneset (e.g. "reactome|tft")}

\item{bg}{numeric background used for functional analysis over-representation test}

\item{dotopnetwork}{logical do top networks}

\item{dotopheatmap}{logical do top heatmap}

\item{layout}{numeric for layout neetwork 1 fr by default 2 dh 3 tree 4 circle 5 grid 6 sphere}

\item{mings}{numeric minimal size of a gene set}

\item{maxgs}{numeric maximal size of a gene set}

\item{overlapmin}{numeric minimal overlap to keep for gene set analysis}

\item{addenarankbarplot}{logical if TRUE add ena barplot ranked by NES score}

\item{dotopgenesetnetwork}{logical do geneset networks}

\item{dotopgenesetheatmap}{logical do geneset heatmap}

\item{dogmtgenesetnetwork}{logical do keyword networks}

\item{dogmtgenesetheatmap}{logical do keyword heatmap}

\item{crosscompint}{logical add cross comparison to results for interaction model}

\item{sample}{numeric analysis using random subset}

\item{seed}{numeric seed for random function}

\item{dopar}{numeric core number}

\item{path}{character results directory path}

\item{dirname}{character results directory name}

\item{zip}{logical compress results directory if TRUE}

\item{remove}{logical remove uncompress results directory if TRUE}
}
\value{
omic results directory
}
\description{
Omic function workflow description:
\itemize{
\item Quality controls and unsupervised analysis: histogram, box plot, PCA and sample clustering.
\item Supervised analysis: analysis of variance (ANOVA) and filter application.
\item Unsupervised analysis for selected features: row clustering, PCA and pattern search across factor levels.
\item Graph generation for selected feature: volcanoplots, heatmaps, lineplots, boxplots, PCA
\item Functional analysis: MSigDB enrichment analysis and STRINGDB interaction network
}

See help("omic") section to test workflow with internal GEO data set GSE65055 and reproduce enrichment results for chromosome cytogenetic bands (doi: 10.1111/cge.12731)
}
\details{
Use moal::env() to load required libraries before moal::omic() (see example)

Use input() function to import and analyse your own data starting from tsv file (or csv with sep = ",")

dat must have one IDs columns in the same order than annotations.

Use annot() function for annotation with Symbol, NCBI, Ensembl IDs.

sif must contains column with description sample corresponding to anova factor analysis.

sif rows must have the same number of samples in the same order that in the dat table.

Experimental design examples for model parameters:
\itemize{
\item 1-way anova: model = "TREATMENT"
\item 2-ways anova: model = "PHENOTYPE+TREATMENT"
\item 2-ways anova with interaction: model = "TREATMENT+TIME+TREATMENT*TIME"
\item 2-ways anova with paired factor: model = "TREATMENT", paired = "CASE"
\item 2-ways anova with batch factor: model = "TREATMENT", batch = "BATCH"
\item 2-ways anova with nested factor: model = "TREATMENT", nested = "CASEinTREATMENT"
\item 3-ways or 4-ways anova (without interaction): model = "PHENOTYPE+TREATMENT+AGE"
}

For paired, batch and nested design, remove batch effect from limma package are used to calculate fold-change

Use dopar = 2 to decrease computing resources.

Use sample for random subset analysis.

To see complete threshold list: moal:::thresholdlist \%>\% lapply("[",c(1,2)) \%>\% unlist \%>\% matrix(ncol=2,byrow = T) \%>\% data.frame \%>\% setNames(c("pval","fc"))

Annotation updates: 22-04-2025 for gene and ensembl, MSigDB 2024.1.Hs, StringDB 12.0
}
\examples{

# # Test workflow with internal GEO data set GSE65055 
# # and reproduce enrichment results for chromosome cytogenetic bands (doi: 10.1111/cge.12731)
# # loading libraries:
# library(moal);moal::env()
# # loading data:
# moal:::GSE65055normdata -> dat
# moal:::GSE65055sampledata -> sif
# # Ordering factors for pairwise comparisons which compute contrast p-values and fold-changes.
# sif$ANEUPLOIDY \%>\% ordered(c("Control","T13","T18","T21")) -> sif$ANEUPLOIDY
# sif$TISSUE \%>\% as.factor -> sif$TISSUE
# # annotation
# dat$rowID \%>\% moal::annot(species= "hs",idtype="GENE",dboutput="ncbi") -> annot
# # omic analysis
# moal::omic(dat,sif,annot,species="hs",model="ANEUPLOIDY",batch="TISSUE",dirname="GSE65055")

}
\author{
Florent Dumont \href{mailto:florent.dumont@universite-paris-saclay.fr}{florent.dumont@universite-paris-saclay.fr}
}
