% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qc.r
\name{qc}
\alias{qc}
\title{Quality Controls}
\usage{
qc(
  dat,
  sif = NULL,
  dooutputinput = FALSE,
  dohisto = TRUE,
  doboxplot = TRUE,
  dohc = TRUE,
  dopca = TRUE,
  breaks = 70,
  dirname = NULL,
  path = "."
)
}
\arguments{
\item{dat}{data.frame first column for rowID column}

\item{sif}{data.frame sample information file}

\item{dooutputinput}{logical if TRUE (by default) export input data}

\item{dohisto}{logical if TRUE (by default) do histogram}

\item{doboxplot}{logical if TRUE (by default) do boxplot}

\item{dohc}{logical if TRUE (by default)do hierarchical clustering}

\item{dopca}{logical if TRUE (by default) do PCA}

\item{breaks}{numeric break number for histogramm function}

\item{dirname}{character}

\item{path}{character}
}
\value{
directory including analysis pdf plots
}
\description{
Descriptive analysis applied on column:
\itemize{
\item histogram, boxplot, hierarchical clustering and PCA for column
}
}
\details{
dat row must be equal to sif row
}
\examples{
# not run
# qc(dat,sif)
}
\author{
Florent Dumont \href{mailto:florent.dumont@universite-paris-saclay.fr}{florent.dumont@universite-paris-saclay.fr}
}
