% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoplot.r
\name{volcanoplot}
\alias{volcanoplot}
\title{Volcanoplot}
\usage{
volcanoplot(
  dat = NULL,
  pval = 0.05,
  fc = 1.5,
  topgenename = TRUE,
  topgenenamen = 5,
  genenamelist = NULL,
  genenamesize = 3,
  title = "Volcanoplot"
)
}
\arguments{
\item{dat}{data.frame table with 4 columns (see details)}

\item{pval}{numeric p-value threshold}

\item{fc}{numeric fold-change threshold}

\item{topgenename}{logical display gene label TRUE by default}

\item{topgenenamen}{numeric increase number of gene label}

\item{genenamelist}{character vector of gene list to label}

\item{genenamesize}{numeric label size for gene name}

\item{title}{character}
}
\value{
no returned value
}
\description{
Do volcanoplot
}
\details{
dat parameter must have 4 columns: rowID , p_AvsB , fc_AvsF and Symbol.
}
\examples{
# not run
# data.frame(rowID,p_AvsB,fc_AvsB,Symbol) -> dat
# volcanoplot(dat)
}
\author{
Florent Dumont \href{mailto:florent.dumont@universite-paris-saclay.fr}{florent.dumont@universite-paris-saclay.fr}
}
