% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volcanoplot.r
\name{volcanoplot}
\alias{volcanoplot}
\title{Volcanoplot}
\usage{
volcanoplot(
  dat,
  pval = 0.05,
  fc = 1.5,
  dogenename = TRUE,
  GeneNameN = 5,
  GeneNameList = NULL,
  GeneNameSize = 2,
  title = "Volcanoplot"
)
}
\arguments{
\item{dat}{data.frame table with 4 columns (see details)}

\item{pval}{numeric p-value threshold}

\item{fc}{numeric fold-change threshold}

\item{dogenename}{logical displya GeneName or not. TRUE by default}

\item{GeneNameN}{logical the number of gene to be displayed. 50 by default}

\item{GeneNameList}{character vector of gene Symbol. see details.}

\item{GeneNameSize}{numeric}

\item{title}{character}
}
\value{
no returned value
}
\description{
Do volcanoplot
}
\details{
dat parameter must have 4 columns: rowID , p_AvsB , fc_AvsF and Symbol.
}
\examples{
# not run
# data.frame(rowID,p_AvsB,fc_AvsB,Symbol) -> dat
# volcanoplot(dat)
}
\author{
Florent Dumont \href{mailto:florent.dumont@universite-paris-saclay.fr}{florent.dumont@universite-paris-saclay.fr}
}
